AGI: Move AgiLoader definitions to loader.h
