+++ engines/agi/agi.h
@@ -541,149 +541,7 @@ struct AgiGame {
 	}
 };
 
-struct AgiDiskVolume {
-	uint32 disk;
-	uint32 offset;
-
-	AgiDiskVolume() : disk(_EMPTY), offset(0) {}
-	AgiDiskVolume(uint32 d, uint32 o) : disk(d), offset(o) {}
-};
-
-/**
- * Apple II version of the format for LOGDIR, VIEWDIR, etc.
- * See AgiLoader_A2::loadDir for more details.
- */
-enum A2DirVersion {
-	A2DirVersionOld,  // 4 bits for volume, 8 for track
-	A2DirVersionNew,  // 5 bits for volume, 7 for track
-};
-
-class AgiLoader {
-public:
-	AgiLoader(AgiEngine *vm) : _vm(vm) {}
-	virtual ~AgiLoader() {}
-
-	/**
-	 * Performs one-time initializations, such as locating files
-	 * with dynamic names.
-	 */
-	virtual void init() {}
-
-	/**
-	 * Loads all AGI directory entries from disk and and populates
-	 * the AgiDir arrays in AgiGame with them.
-	 */
-	virtual int loadDirs() = 0;
-
-	/**
-	 * Loads a volume resource from disk.
-	 */
-	virtual uint8 *loadVolumeResource(AgiDir *agid) = 0;
-
-	/**
-	 * Loads AgiEngine::_objects from disk.
-	 */
-	virtual int loadObjects() = 0;
-
-	/**
-	 * Loads AgiBase::_words from disk.
-	 */
-	virtual int loadWords() = 0;
-
-protected:
-	AgiEngine *_vm;
-};
-
-class AgiLoader_A2 : public AgiLoader {
-public:
-	AgiLoader_A2(AgiEngine *vm) : AgiLoader(vm) {}
-	~AgiLoader_A2() override;
-
-	void init() override;
-	int loadDirs() override;
-	uint8 *loadVolumeResource(AgiDir *agid) override;
-	int loadObjects() override;
-	int loadWords() override;
-
-private:
-	Common::Array<Common::SeekableReadStream *> _disks;
-	Common::Array<AgiDiskVolume> _volumes;
-	AgiDir _logDir;
-	AgiDir _picDir;
-	AgiDir _viewDir;
-	AgiDir _soundDir;
-	AgiDir _objects;
-	AgiDir _words;
-
-	int readDiskOne(Common::SeekableReadStream &stream, Common::Array<uint32> &volumeMap);
-	static bool readInitDir(Common::SeekableReadStream &stream, byte index, AgiDir &agid);
-	static bool readDir(Common::SeekableReadStream &stream, int position, AgiDir &agid);
-	static bool readVolumeMap(Common::SeekableReadStream &stream, uint32 position, uint32 bufferLength, Common::Array<uint32> &volumeMap);
-
-	A2DirVersion detectDirVersion(Common::SeekableReadStream &stream);
-	bool loadDir(AgiDir *dir, Common::SeekableReadStream &disk, uint32 dirOffset, uint32 dirLength, A2DirVersion dirVersion);
-};
-
-class AgiLoader_v1 : public AgiLoader {
-public:
-	AgiLoader_v1(AgiEngine *vm) : AgiLoader(vm) {}
-
-	void init() override;
-	int loadDirs() override;
-	uint8 *loadVolumeResource(AgiDir *agid) override;
-	int loadObjects() override;
-	int loadWords() override;
-
-private:
-	Common::Array<Common::String> _imageFiles;
-	Common::Array<AgiDiskVolume> _volumes;
-	AgiDir _logDir;
-	AgiDir _picDir;
-	AgiDir _viewDir;
-	AgiDir _soundDir;
-	AgiDir _objects;
-	AgiDir _words;
-
-	bool readDiskOneV1(Common::SeekableReadStream &stream);
-	bool readDiskOneV2001(Common::SeekableReadStream &stream, int &vol0Offset);
-	static bool readInitDirV1(Common::SeekableReadStream &stream, byte index, AgiDir &agid);
-	static bool readInitDirV2001(Common::SeekableReadStream &stream, byte index, AgiDir &agid);
-
-	bool loadDir(AgiDir *dir, Common::File &disk, uint32 dirOffset, uint32 dirLength);
-};
-
-class AgiLoader_v2 : public AgiLoader {
-private:
-	bool _hasV3VolumeFormat;
-
-	int loadDir(AgiDir *agid, const char *fname);
-	bool detectV3VolumeFormat();
-
-public:
-	AgiLoader_v2(AgiEngine *vm) : _hasV3VolumeFormat(false), AgiLoader(vm) {}
-
-	int loadDirs() override;
-	uint8 *loadVolumeResource(AgiDir *agid) override;
-	int loadObjects() override;
-	int loadWords() override;
-};
-
-class AgiLoader_v3 : public AgiLoader {
-private:
-	Common::String _name; /**< prefix in directory and/or volume file names (`GR' for goldrush) */
-
-	int loadDir(AgiDir *agid, Common::File *fp, uint32 offs, uint32 len);
-
-public:
-	AgiLoader_v3(AgiEngine *vm) : AgiLoader(vm) {}
-
-	void init() override;
-	int loadDirs() override;
-	uint8 *loadVolumeResource(AgiDir *agid) override;
-	int loadObjects() override;
-	int loadWords() override;
-};
-
+class AgiLoader;
 class GfxFont;
 class GfxMgr;
 class SpritesMgr;
