+++ engines/agi/loader_v2.cpp
@@ -22,6 +22,7 @@
 #include "common/textconsole.h"
 
 #include "agi/agi.h"
+#include "agi/loader.h"
 #include "agi/lzw.h"
 #include "agi/words.h"
